02-langchain-llamaindex/requirements.txt
openai>=2.15.0
python-dotenv>=1.2.1
jsonschema>=4.26.0
langchain>=1.2.7
langchain-openai>=1.1.7
llama-index>=0.14.13
llama-index-llms-openai>=0.6.13